requirements.txt
annotated-types==0.7.0
anyio==4.9.0
arxiv==2.2.0
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
biopython==1.85
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
chromadb==1.0.13
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
concurrent-log-handler==0.9.24
decorator==5.2.1
distro==1.9.0
durationpy==0.10
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.5.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
langchain==0.3.25
langchain-core==0.3.65
langchain-text-splitters==0.3.8
langsmith==0.3.45
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==5.1.0
moviepy==2.2.1
mpmath==1.3.0
networkx==3.5
newspaper3k
numpy==2.3.0
oauthlib==3.3.0
onnxruntime==1.22.0
openai==1.88.0
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.10.18
overrides==7.7.0
packaging==24.2
Pillow==11.2.1
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
pip==25.1.1
pluggy==1.6.0
posthog==5.3.0
proglog==0.1.12
protobuf==5.29.5
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybase64==1.4.1
pydantic==2.11.7
pydantic-core==2.33.2
Pygments==2.19.1
PyMuPDF==1.26.1
PyPika==0.48.9
pyproject-hooks==1.2.0
pyreadline3==3.5.4
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
tenacity==8.2.3
rich==14.0.0
rsa==4.9.1
rpds-py==0.25.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
selenium==4.22.0
sentence-transformers==4.1.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.1
tqdm==4.67.1
trafilatura
transformers==4.52.4
typer==0.16.0
typing-extensions==4.14.0
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.34.3
watchfiles==1.1.0
webdriver-manager==4.0.1
websocket-client==1.8.0
websockets==15.0.1
whisper==1.1.10
zipp==3.23.0
zstandard==0.23.0
readability-lxml==0.8.1
boilerpy3==1.0.6
lxml